refactored map-overview into ui folder in tg-frontend
