+++ tg-frontend/src/widgets/map-overview/ui/__mocks.ts
