+++ tg-frontend/src/widgets/map-overview/ui/map-overview.stories.tsx
